+++ src/main/java/com/yeohangttukttak/api/domain/place/Place.java
@@ -3,14 +3,12 @@ package com.yeohangttukttak.api.domain.place;
 import com.yeohangttukttak.api.domain.Attachable;
 import com.yeohangttukttak.api.domain.BaseEntity;
 import com.yeohangttukttak.api.domain.file.File;
-import com.yeohangttukttak.api.domain.travel.*;
+import com.yeohangttukttak.api.domain.visit.entity.Visit;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DialectOverride;
 import org.hibernate.annotations.SQLRestriction;
-import org.hibernate.annotations.Where;
 import org.locationtech.jts.geom.Point;
 
 import java.util.ArrayList;
