+++ src/test/java/com/yeohangttukttak/api/repository/VisitRepositoryTest.java
@@ -3,7 +3,8 @@ package com.yeohangttukttak.api.repository;
 import com.yeohangttukttak.api.domain.place.Location;
 import com.yeohangttukttak.api.domain.place.Place;
 import com.yeohangttukttak.api.domain.travel.Travel;
-import com.yeohangttukttak.api.domain.travel.Visit;
+import com.yeohangttukttak.api.domain.visit.entity.Visit;
+import com.yeohangttukttak.api.domain.visit.dao.VisitRepository;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
